added Exchange factory
