+++ univocity-trader-core/src/main/java/com/univocity/trader/ClientAccount.java
@@ -2,6 +2,7 @@ package com.univocity.trader;
 
 import com.univocity.trader.account.*;
 import com.univocity.trader.candles.*;
+import com.univocity.trader.exchange.Exchange;
 
 import java.util.*;
