+++ univocity-trader-core/src/main/java/com/univocity/trader/SymbolPriceDetails.java
@@ -1,6 +1,7 @@
 package com.univocity.trader;
 
 import com.univocity.trader.candles.*;
+import com.univocity.trader.exchange.Exchange;
 
 import java.math.*;
 import java.util.*;
