+++ univocity-trader-core/src/main/java/com/univocity/trader/account/OrderManager.java
@@ -3,6 +3,7 @@ package com.univocity.trader.account;
 
 import com.univocity.trader.*;
 import com.univocity.trader.candles.*;
+import com.univocity.trader.exchange.Exchange;
 import com.univocity.trader.indicators.*;
 import com.univocity.trader.indicators.base.*;
 import com.univocity.trader.strategy.*;
