+++ univocity-trader-core/src/main/java/com/univocity/trader/account/TradingManager.java
@@ -2,6 +2,7 @@ package com.univocity.trader.account;
 
 import com.univocity.trader.*;
 import com.univocity.trader.candles.*;
+import com.univocity.trader.exchange.Exchange;
 import com.univocity.trader.indicators.base.*;
 import com.univocity.trader.notification.*;
 import com.univocity.trader.simulation.*;
