+++ univocity-trader-core/src/main/java/com/univocity/trader/simulation/SimulatedExchange.java
@@ -3,6 +3,7 @@ package com.univocity.trader.simulation;
 import com.univocity.trader.*;
 import com.univocity.trader.account.*;
 import com.univocity.trader.candles.*;
+import com.univocity.trader.exchange.Exchange;
 import com.univocity.trader.indicators.base.*;
 
 import java.util.*;
