+++ univocity-trader-marketsimulator/pom.xml
@@ -17,11 +17,6 @@
 	<dependencies>
 		<dependency>
 			<groupId>com.univocity</groupId>
-			<artifactId>univocity-trader-binance</artifactId>
-			<version>${revision}</version>
-		</dependency>
-		<dependency>
-			<groupId>com.univocity</groupId>
 			<artifactId>univocity-trader-examplestrategy</artifactId>
 			<version>${revision}</version>
 		</dependency>
