+++ univocity.properties
@@ -15,5 +15,6 @@ mail.password=<YOUR SMTP PASSWORD>
 mail.sender=<YOU>>@gmail.com
 
 # exchange
+exchange.class=com.univocity.trader.exchange.binance.BinanceExchange
 exchange.apikey=<YOUR BINANCE API KEY>"
 exchange.secret=<YOUR BINANCE API SECRET>"
